chore!: upgrade to cdktf 0.18.0 (#202)
This PR initiates the upgrade of CDKTF from version `0.17.0` to version
`0.18.0`.
Unfortunately, not everything can be automated, and the following steps
need to be completed manually:
 
- [x] Update `@cdktf/provider-kubernetes` to a version compatible with
`cdktf@0.18.0`
[here](https://github.com/cdktf/cdktf-cdk8s/blob/4d23de9f0dadddb922bf1c44079f08856491f11b/.projenrc.ts#L67).
Look up the version
[here](https://github.com/cdktf/cdktf-provider-kubernetes/releases/).
- [x] Run `npx projen`
 
Please checkout this PR, complete the above steps, push the changes to
this branch, and then mark this PR as ready for review to complete the
upgrade. Thanks!

---------

Signed-off-by: team-tf-cdk <github-team-tf-cdk@hashicorp.com>
Co-authored-by: Nara Kasbergen <nara.kasbergen@gmail.com>