chore(deps): upgrade dependencies in examples (#328) [skip ci]
We're getting a lot of security warnings about dependencies in the
examples. Dependabot is set up to manage the examples, but only for
security updates, and because of all of the complex interdependencies,
in practice it's not set up to fix issues most of the time. This
manually upgrades all of the examples on a one-time basis, and sets up
Dependabot to manage these dependencies on a weekly basis going forward.